requirements.txt
aiosmtplib==2.0.2
annotated-types==0.5.0
anyio==4.0.0
bcrypt==4.0.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.103.0
greenlet==2.0.2
h11==0.14.0
httptools==0.6.0
idna==3.4
Jinja2==3.1.2
libmagic==1.0
MarkupSafe==2.1.3
passlib==1.7.4
protobuf==4.24.3
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
pyotp==2.9.0
pypng==0.20220715.0
python-dotenv==1.0.0
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.6
PyYAML==6.0.1
qrcode==7.4.2
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
